feat(dashboard): move AI scan targets into the Device Agent configuration tab
