+++ .changeset/ai-scan-targets-dashboard.md
@@ -2,6 +2,6 @@
 "dashboard": minor
 ---
 
-feat: add the AI Scan Targets page under Platform Admin
+feat: manage the Shadow AI scan target catalog from the Device Agent configuration tab
 
-Platform administrators can list, add, edit, enable, disable, and delete the Shadow AI scan targets served to every enrolled device agent, and see the catalog's recent revisions. The strict platform-admin gate used by OpenRouter Keys is now a shared component.
+Platform administrators can list, add, edit, enable, disable, and delete the Shadow AI scan targets served to every enrolled device agent from a new section on the Device Agent Configuration tab. The strict platform-admin gate used by OpenRouter Keys is now a shared component.
