+++ client/dashboard/src/components/org-sidebar.tsx
@@ -114,7 +114,6 @@ export function OrgSidebar({
     orgRoutes.platformAdminRbac,
     orgRoutes.platformAdminOnboarding,
     orgRoutes.platformAdminOpenRouterKeys,
-    orgRoutes.platformAdminAiScanTargets,
     orgRoutes.platformRemoteIdentityProviders,
   ].some((r) => r.active);
 
@@ -153,7 +152,6 @@ export function OrgSidebar({
     orgRoutes.platformAdminRbac,
     orgRoutes.platformAdminOnboarding,
     orgRoutes.platformAdminOpenRouterKeys,
-    orgRoutes.platformAdminAiScanTargets,
     orgRoutes.platformRemoteIdentityProviders,
   ];
   const activeRoute = allOrgNavRoutes.find((r) => r.active);
@@ -315,18 +313,13 @@ export function OrgSidebar({
                     : []),
                   ...(isPlatformAdmin
                     ? [
-                        // OpenRouter Keys, AI Scan Targets and Remote Identity
-                        // Providers stay strictly admin-gated even in local
-                        // dev: all manage real platform state (live upstream
-                        // credentials, the catalog every device probes with,
-                        // the shared issuer catalog), not local developer aids.
+                        // OpenRouter Keys and Remote Identity Providers stay
+                        // strictly admin-gated even in local dev: both manage
+                        // real platform state (live upstream credentials, the
+                        // shared issuer catalog), not local developer aids.
                         {
                           item: orgRoutes.platformAdminOpenRouterKeys,
                           label: "OpenRouter Keys",
-                        },
-                        {
-                          item: orgRoutes.platformAdminAiScanTargets,
-                          label: "AI Scan Targets",
                         },
                         {
                           item: orgRoutes.platformRemoteIdentityProviders,
