+++ client/dashboard/src/pages/device-agent/ai-scan-target-draft.ts
